Staging

+++ app/Http/Controllers/TaskController.php
@@ -38,21 +38,30 @@ class TaskController extends Controller
         // Fetch tasks only related to the currently active fortnight if currentFortnight is true
         $currentFortnight = null;
 
+        $today = Carbon::now()->format('Y-m-d');
         if ($request->query('currentFortnight')) {
 
-            $today = Carbon::now()->format('Y-m-d');
+
+            $currentFortnight = Fortnight::whereDate('start_date', '<=', $today)
+                ->whereDate('end_date', '>=', $today)->first();
+                
+                $tasks = $tasks->whereHas('fortnights', function ($query) use ($currentFortnight) {
+                    $query->where('fortnights.id', $currentFortnight->id);
+                });
+        } elseif ($request->query('todays')) {
 
             $currentFortnight = Fortnight::whereDate('start_date', '<=', $today)
                 ->whereDate('end_date', '>=', $today)->first();
 
-            $tasks = $tasks->whereHas('fortnights', function ($query) use ($currentFortnight) {
-                $query->where('fortnights.id', $currentFortnight->id);
-            });
-        } elseif ($request->query('todays')) {
-            
-            $today = Carbon::now()->format('Y-m-d');
-
-            $date = Day::where('date', $today)->first()->id;
+            if (Day::where('date', $today)->exists()) {
+                $date = Day::where('date', $today)->first();
+            } else {
+                $date = Day::create([
+                        'fortnight_id' => $currentFortnight->id, 
+                        'date' => $today
+                    ]);
+            }
+    
 
             $tasks = $tasks->whereHas('days', function ($query) use ($date) {
                 $query->where('days.id', $date);
@@ -139,13 +148,14 @@ class TaskController extends Controller
             $today = Carbon::now()->format('Y-m-d');
             $currentFortnightId = Fortnight::whereDate('start_date', '<=', $today)
                 ->whereDate('end_date', '>=', $today)->first()->id;
-    
-            $day = Day::firstOrCreate([
-                'date' => $request->query('today')], 
-                [
-                    'fortnight_id' => $currentFortnightId, 
-                    'date' => $today
-                ]);
+            if (Day::where('date', $today)->exists()) {
+                $day = Day::where('date', $today)->first();
+            } else {
+                $day = Day::create([
+                        'fortnight_id' => $currentFortnightId, 
+                        'date' => $today
+                    ]);
+            }
 
             $day->tasks()->attach($task);
         }
@@ -177,6 +187,7 @@ class TaskController extends Controller
         $task = $task->load('days');
 
         $isDaily = $task->days()->exists();
+
         $today = $isDaily ? $task->days()->first() : null;
 
         $assignedUsers = $task->users()->pluck('users.id')->toArray();
@@ -202,28 +213,32 @@ class TaskController extends Controller
 
         $data = $request->validated();
 
-
         $data['is_subtask'] = $data['parent_task_id'] ? true : false;
 
         $departments = $request['department_id'];
         $fortnights = $request['fortnight_id'];
-        $users = $request['user_id'];
+        $users = request()->user()->hasRole('EMPLOYEE') ? [0 => request()->user()->id] : $request['user_id'];
 
-        $task->departments()->detach();
+        if (!request()->user()->hasRole('EMPLOYEE')) {
+            $task->departments()->detach();
+            $task->users()->detach();
+        }
+
         $task->fortnights()->detach();
-        $task->users()->detach();
 
         unset($data['department_id']);
         unset($data['fortnight_id']);
         unset($data['user_id']);
 
         $task->update($data);
+
         if (!request()->user()->hasRole('EMPLOYEE')) {
             $task->departments()->attach($departments);
+            $task->fortnights()->attach($fortnights);
             $task->users()->attach($users);
-            # code...
         }
 
+        dd($data);
         return redirect()->route('tasks.index')->with('status', 'Task has been successfully Updated.');
     }
 
@@ -235,7 +250,7 @@ class TaskController extends Controller
 
         if ($task->kpis()->exists() || $task->deliverables()->exists() || $task->feedbacks()->exists()) {
             return redirect()->route('tasks.index')
-                ->with('related', 'task-deleted');
+            ->with('status', 'You can\'t Delete This pending Task it have feedback on it.');
         }
 
         $task->delete();
